+++ docs/html/search/all_1.js
@@ -1,6 +1,4 @@
 var searchData=
 [
-  ['add_20library_20to_20your_20project_0',['Add library to your project',['../index.html#autotoc_md2',1,'']]],
-  ['aux_1',['aux',['../group__timer0__to__ms.html',1,'Timer0 Aux'],['../group__timer2__to__ms.html',1,'Timer2 Aux']]],
-  ['aux_20common_2',['Timer Aux common',['../group__timer__to__ms.html',1,'']]]
+  ['_5f_5fat_0',['__at',['../group__chip__id.html#gaf797f652068877217b527fa8d320bb43',1,'chip_id.h']]]
 ];