+++ docs/html/search/all_b.js
@@ -1,4 +1,16 @@
 var searchData=
 [
-  ['none_0',['NONE',['../group__freq.html#ggae5d505beabdbd466d648cb5dccf4b6d2ac157bdf0b85a40d2619cbc8bc1ae5fe2',1,'frequency.h']]]
+  ['main_0',['main',['../counter0__mode1__example_8c.html#acdef7a1fd863a6d3770c1268cb06add3',1,'counter0_mode1_example.c']]],
+  ['main_5ffosc_1',['MAIN_Fosc',['../sys_8h.html#aa52cf4efd75d43c7917d9eec3becd2fa',1,'sys.h']]],
+  ['main_5ffosch_2',['MAIN_FoscH',['../sys_8h.html#ae1b9b7f30c88b5491fd35d286418f6d6',1,'sys.h']]],
+  ['main_5ffoscl_3',['MAIN_FoscL',['../sys_8h.html#ab42e14bee431f5da3ecef889588cc728',1,'sys.h']]],
+  ['management_4',['Power management',['../group__power__management.html',1,'']]],
+  ['master_5fclock_5foutput_5fpin_5ft_5',['master_clock_output_pin_t',['../group__freq.html#gae5d505beabdbd466d648cb5dccf4b6d2',1,'frequency.h']]],
+  ['maturity_6',['Library modules maturity',['../index.html#autotoc_md5',1,'']]],
+  ['mode_200_7',['UART1 Mode 0',['../group__uart1__mode0.html',1,'']]],
+  ['mode0_8',['mode0',['../group__counter0__mode0.html',1,'Counter0 Mode0'],['../group__counter2__mode0.html',1,'Counter2 Mode0'],['../group__timer0__mode0.html',1,'Timer0 Mode0'],['../group__timer2__mode0.html',1,'Timer2 Mode0']]],
+  ['mode1_9',['mode1',['../group__counter0__mode1.html',1,'Counter0 Mode1'],['../group__timer0__mode1.html',1,'Timer0 Mode1']]],
+  ['mode2_10',['mode2',['../group__counter0__mode2.html',1,'Counter0 Mode2'],['../group__timer0__mode2.html',1,'Timer0 Mode2']]],
+  ['mode3_11',['mode3',['../group__counter0__mode3.html',1,'Counter0 Mode3'],['../group__timer0__mode3.html',1,'Timer0 Mode3']]],
+  ['modules_20maturity_12',['Library modules maturity',['../index.html#autotoc_md5',1,'']]]
 ];