+++ docs/html/search/files_8.js
@@ -1,4 +1,5 @@
 var searchData=
 [
-  ['uart_2eh_0',['uart.h',['../uart_8h.html',1,'']]]
+  ['uart_2eh_0',['uart.h',['../uart_8h.html',1,'']]],
+  ['uart1_5fmode0_2eh_1',['uart1_mode0.h',['../uart1__mode0_8h.html',1,'']]]
 ];