+++ docs/html/search/groups_0.js
@@ -1,5 +1,4 @@
 var searchData=
 [
-  ['aux_0',['aux',['../group__timer0__to__ms.html',1,'Timer0 Aux'],['../group__timer2__to__ms.html',1,'Timer2 Aux']]],
-  ['aux_20common_1',['Timer Aux common',['../group__timer__to__ms.html',1,'']]]
+  ['0_0',['UART1 Mode 0',['../group__uart1__mode0.html',1,'']]]
 ];