+++ docs/html/search/groups_a.js
@@ -1,4 +1,5 @@
 var searchData=
 [
-  ['watchdog_20timer_0',['Watchdog timer',['../group__wdt.html',1,'']]]
+  ['uart1_0',['UART1',['../group__uart.html',1,'']]],
+  ['uart1_20mode_200_1',['UART1 Mode 0',['../group__uart1__mode0.html',1,'']]]
 ];