+++ docs/html/search/searchdata.js
@@ -1,12 +1,12 @@
 var indexSectionsWithContent =
 {
-  0: "_acdefghilmnoprstuvwy",
+  0: "0_acdefghilmnoprstuvwy",
   1: "cdfiprstuw",
   2: "_cgmpu",
   3: "eimt",
   4: "hlnoprt",
   5: "clm",
-  6: "acdfimprtuw",
+  6: "0acdfimprtuw",
   7: "hivw"
 };