+++ src/uart1_mode0.c
@@ -1,0 +1,6 @@
+#include <uart1_mode0.h>
+
+void uart1_mode0_init()
+{
+    
+}